requirements.txt
cirq
tensorboard==2.0.0
torchsummary
pathos